rotate macs, fix setting env and starting wg

+++ wireguard/del.sh
@@ -10,9 +10,14 @@ if grep -q "$1" <<<"$peers"; then
     sed -i "s/$peers/$(sed "s/,\?$1//" <<<"$peers")/" compose.yml
     sudo rm -rf wireguard/config/peer_"$1"
     sudo mv -f wireguard/config/wg_confs/wg0.conf wireguard/config/wg_confs/wg0.conf.bak
-    sudo docker compose restart wireguard
-    sudo docker exec wireguard bash -c "wg-quick down wg0 ; wg-quick up wg0"
-    sudo docker compose up -d wireguard
+
+    if $CLS_DOCKER; then
+        sudo docker compose restart wireguard
+        sudo docker exec wireguard bash -c "wg-quick down wg0 ; wg-quick up wg0"
+        sudo docker compose up -d wireguard
+    else
+        sudo wg-quick down "$CLS_INTERN_IFACE" ; sudo wg-quick up "$CLS_INTERN_IFACE"
+    fi
 fi
 
 popd &>/dev/null || exit
